requirements.txt

raphlib
networkx
plotly
streamlit_plotly_events
